test: faster init for sweat storage stress-test
Use `prepare_accounts` (introduced in #9428)
to create the 1000 users required to start
the sweat based storage stress test.

Also, we don't have to send FT balance to those users,
or even add a key to the oracle account. Just creating the
account and sending some storage balance is good enough.

If running with 50 users and ramping up 1 user/sec,
this reduced the initialization time of the stress test
from 3m20s to 1m30s. Not perfect but better.

+++ pytest/tests/loadtest/locust/common/base.py
@@ -188,7 +188,7 @@ class Deploy(Transaction):
 
 class CreateSubAccount(Transaction):
 
-    def __init__(self, sender, sub_key, balance: int = 50):
+    def __init__(self, sender, sub_key, balance: float = 50):
         super().__init__()
         self.sender = sender
         self.sub_key = sub_key
@@ -358,7 +358,7 @@ class NearNodeProxy:
     def account_exists(self, account_id: str) -> bool:
         return "error" not in self.node.get_account(account_id, do_assert=False)
 
-    def prepare_account(self, account: Account, parent: Account, balance: int,
+    def prepare_account(self, account: Account, parent: Account, balance: float,
                         msg: str) -> bool:
         """
         Creates the account if it doesn't exist and refreshes the nonce.
@@ -373,7 +373,7 @@ class NearNodeProxy:
     def prepare_accounts(self,
                          accounts: typing.List[Account],
                          parent: Account,
-                         balance: int,
+                         balance: float,
                          msg: str,
                          timeout: timedelta = timedelta(minutes=3)):
         """
